+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## UNRELEASED
+
+## `reitit-ring`
+
+* router is injected into request also in the default branch
+
+## `reitit-http`
+
+* router is injected into request also in the default branch
+
 ## 0.2.4 (21.10.2018)
 
 ## `reitit-ring`